add conveyor belt code

+++ gameserver/src/main/java/brainwine/gameserver/server/messages/PlayerPositionMessage.java
@@ -1,15 +1,15 @@
 package brainwine.gameserver.server.messages;
 
-import brainwine.gameserver.server.Message;
 import brainwine.gameserver.server.MessageInfo;
+import brainwine.gameserver.server.Message;
 
 @MessageInfo(id = 5)
 public class PlayerPositionMessage extends Message {
     
-    public int x;
-    public int y;
+    public float x;
+    public float y;
     
-    public PlayerPositionMessage(int x, int y) {
+    public PlayerPositionMessage(float x, float y) {
         this.x = x;
         this.y = y;
     }
